requirements.txt
# ChromaDB
chromadb>=0.4.0

# Web framework
flask>=3.0.0

# Google embeddings (same as main bot)
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Document processing
pypdf>=3.0.0
python-docx>=1.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
html2text>=2020.1.16
lxml>=5.0.0

# AI summarization
anthropic>=0.40.0

# Utilities
requests>=2.31.0

# Production server
gunicorn>=21.0.0
